% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_stats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{GET Request for Player Statistics}
\usage{
get_player_stats(week, season, league_id)
}
\arguments{
\item{week}{The week you wish to get stats for. If blank all availble statistics for the season chosen will be returned}

\item{season}{The year of the season you wish to get stats for}

\item{league_id}{The ID of you ESPN fantasy league}
}
\value{
A dataframe containing player projections and actual game statistics for the timeframe chosen.
}
\description{
A function for accessing NFL player statistics via the Coinbase API.
}
